+++ core/MemoryBus.h
@@ -10,10 +10,10 @@ public:
     MemoryBus(CPU &cpu);
     void reset();
 
-    uint8_t read(uint16_t addr) const;
-    void write(uint16_t addr, uint8_t data);
+    uint8_t read(uint32_t addr) const;
+    void write(uint32_t addr, uint8_t data);
 
-    const uint8_t *mapAddress(uint16_t addr) const;
+    const uint8_t *mapAddress(uint32_t addr) const;
 
 private:
